Misc optimizations

+++ src/graysvr/CItem.h
@@ -895,7 +895,7 @@ public:
 	bool OnSpellEffect(SPELL_TYPE spell, CChar *pCharSrc, int iSkillLevel, CItem *pSourceItem, bool bReflecting = false);
 	int OnTakeDamage(int iDmg, CChar *pSrc, DAMAGE_TYPE uType = DAMAGE_HIT_BLUNT);
 
-	int Armor_GetRepairPercent() const;
+	WORD Armor_GetRepairPercent() const;
 	LPCTSTR Armor_GetRepairDesc() const;
 	bool Armor_IsRepairable() const;
 	int Armor_GetDefense() const;
